requirements_base.txt
otree-core>=1.3.9
otreechat
Django==1.8.8 # for heroku, needs to be explicitly in requirements file
